fixed long-standing bug in how slice position is parsed on 64-bit machines (thanks Christoph Garth!)
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@3903 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/unrrdu/slice.c
@@ -36,7 +36,7 @@ unrrdu_sliceMain(int argc, char **argv, char *me, hestParm *hparm) {
   unsigned int axis;
   int pret;
   size_t pos;
-  int _pos[2];
+  long int _pos[2];
   airArray *mop;
 
   OPT_ADD_AXIS(axis, "axis to slice along");
